Small fixes on Mon Sep 15 (#16)
* Fixing some printing issues and adding a method to change term name

* Forgot to save changes

* Addressing warnings in tests

* Correcting print and valgrind

* n_dupl_events and n_spec_events are not functions

* Replacing uint with size_t everywhere

* Reversing changes in catch

* Handling nullptr print of the BArray::print method

* Removing trash

* Addressing comments by copilot and updating copilot-instructions.md

* Updating version number

* Minor commit correcting the deploy action

+++ defm.hpp
@@ -160,7 +160,11 @@ inline void DEFMData::print() const {
     for (size_t i = 0u; i < array->nrow(); ++i)
     {
 
-        printf_barry("row %li (%li): ", i, obs_start + i);
+        printf_barry(
+            "row %i (%i): ",
+            static_cast<int>(i),
+            static_cast<int>(obs_start + i)
+        );
         for (size_t j = 0u; j < X_ncol; ++j)
         {
             printf_barry("% 5.2f, ", operator()(i, j));
@@ -287,12 +291,13 @@ typedef barry::Rules<DEFMArray, DEFMRuleDynData> DEFMRulesDyn;
  * which we call LHS, can only hold `row id` that are less than `m_order`.
  * 
  * 
- * 
- * @param formula 
- * @param locations 
- * @param signs 
- * @param m_order 
- * @param y_ncol 
+ * @param formula A string specifying the motif formula (see details).
+ * @param locations A vector of locations for the motif variables.
+ * @param signs A vector of signs for the motif variables.
+ * @param m_order The Markov order.
+ * @param y_ncol The number of columns in the response variable.
+ * @param covar_name A string to hold the name of the covariate (if any).
+ * @param vname A string to hold the variable name (if any).
  */
 inline void defm_motif_parser(
     std::string formula,
@@ -1567,7 +1572,6 @@ inline void DEFM::simulate(
                 for (size_t y = 0u; y < Y_ncol; ++y)
                     *(y_out + n_entry++) = last_array(M_order, y, false);
 
-                // last_array.print("i: %li, proc_n: %li\n", i, proc_n);
 
             }
             else
@@ -1588,16 +1592,10 @@ inline void DEFM::simulate(
                     true // Delete the data
                 );
 
-                // Baseline
-                // tmp_array.print("baseline i: %li, proc_n: %li\n", i, proc_n);
-                // tmp_array.D().print();
-
                 model_num++;
                 last_array = this->sample(tmp_array, par);
                 for (size_t y = 0u; y < Y_ncol; ++y)
                     *(y_out + n_entry++) = last_array(M_order, y, false);
-
-                // last_array.print("generated i: %li, proc_n: %li\n", i, proc_n);
 
             }
